[CST] add library for cst
