+++ Cargo.toml
@@ -1,3 +1,3 @@
 [workspace]
 resolver = "3"
-members = ["pp", "location", "error"]
+members = ["pp", "location", "error", "cst"]
