+++ cst/src/file.rs
@@ -1,7 +1,6 @@
 use super::Command;
-use crate::lexer::meta::Meta;
-use crate::utils::pretty::Pretty;
-use crate::utils::theme::{Doc, Theme};
+use crate::Meta;
+use pp::prelude::*;
 
 // ============================================================================
 // EndOf File
