+++ cst/src/lib.rs
@@ -1,9 +1,11 @@
 pub mod command;
+pub mod comment;
 pub mod constant;
 pub mod expression;
 pub mod expression_definition;
 pub mod file;
 pub mod help;
+pub mod meta;
 pub mod operator;
 pub mod parenthesis;
 pub mod pattern;
@@ -12,11 +14,13 @@ pub mod ty_definition;
 
 pub use command::Command;
 pub use command::CommandKind;
+pub use comment::Comment;
 pub use constant::Constant;
 pub use expression::Expression;
 pub use expression_definition::ExpressionDefinition;
 pub use file::EndOfFile;
 pub use file::File;
+pub use meta::Meta;
 pub use pattern::Pattern;
 pub use ty::Type;
 pub use ty_definition::TypeDefinition;
