+++ pp/src/lib.rs
@@ -1,2 +1,8 @@
-pub mod theme;
 pub mod pretty;
+pub mod theme;
+
+pub mod prelude {
+    pub use crate::pretty::Pretty;
+    pub use crate::theme::Doc;
+    pub use crate::theme::Theme;
+}
