+++ pp/src/pretty.rs
@@ -23,6 +23,39 @@ pub trait Pretty: Sized {
     }
 }
 
+// ============================================================================
+// Pretty with Precedence
+// ============================================================================
+pub trait PrettyPrecedence {
+    /// get level of type
+    fn precedence(&self) -> u8;
+
+    /// pretty with precedence
+    fn pretty_precedence(&self, min_prec: u8, theme: &Theme) -> Doc<'_>;
+}
+
+impl<T> Pretty for T
+where
+    T: PrettyPrecedence,
+{
+    fn pretty(&self, theme: &Theme) -> Doc<'_> {
+        self.pretty_precedence(self.precedence(), theme)
+    }
+}
+
+impl<T> PrettyPrecedence for Box<T>
+where
+    T: PrettyPrecedence,
+{
+    fn precedence(&self) -> u8 {
+        self.as_ref().precedence()
+    }
+
+    fn pretty_precedence(&self, min_prec: u8, theme: &Theme) -> Doc<'_> {
+        self.as_ref().pretty_precedence(min_prec, theme)
+    }
+}
+
 // ===========================================================================
 // Stream Colored
 // ===========================================================================
