+++ src/Cargo.toml
@@ -1,0 +1,30 @@
+[package]
+name = "startlang"
+version = "0.0.1"
+edition = "2021"
+
+[build-dependencies]
+cc="*"
+glob = "0.3.3"
+
+[dev-dependencies]
+trycmd = "0.15.10"
+assert_cmd = "2"
+serde_json = "1.0.142"
+assert-json-diff = "2.0.2"
+temp-dir = "0.1.16"
+fs_extra = "1.3.0"
+serde = "1.0.219"
+
+[dependencies]
+ariadne = { version = "0.5.1", features = ["auto-color"] }
+chumsky = "0.10.1"
+clap = { version = "4.5.45", features = ["derive"] }
+color-print = "0.3.7"
+colored = "3.0.0"
+num-bigint = "0.4.6"
+pretty = "0.12.4"
+rustyline = "17.0.0"
+similar = "2.7.0"
+tokio = { version = "1", features = ["full"] }
+tower-lsp = "0.20.0"
